Stage 6: quality gates
ruff:
- Removed deprecated ANN101/ANN102 ignores; added BLE to selected
  rules so blind-except patterns are flagged.
- Replaced try/except/pass in cli._ensure_utf8_streams with
  contextlib.suppress(Exception).
- Switched the single intentional blind-except (top-level CLI
  boundary mapping all errors to exit codes) to a documented
  noqa: BLE001.
- Replaced int(round(...)) with plain round() in
  NtpTimeSource and its test (RUF046).
- Tightened test pytest.raises calls with `match=` patterns and raw
  strings where they contained regex metacharacters (PT011, RUF043).
- Sorted domain/__init__.py __all__ (RUF022).
- Replaced an EN DASH in the Slovenian description with a hyphen
  (RUF001) — Fisher-Yates with ASCII hyphen.
- Annotated the synthetic test PKCS#12 password with noqa: S105.
- Ran ruff format across src/ and tests/ — 17 files reformatted,
  imports auto-sorted in test_end_to_end.py.

mypy --strict: still clean across 18 source files.

Coverage:
- pyproject.toml exclude_lines now skips Protocol method `...` bodies.
- Marked Protocol method bodies in application/ports.py and
  domain/randomizer.py with `# pragma: no cover` so the coverage
  report no longer reports phantom 79% on ports.py.
- Domain and application both at 100% coverage; total 96.60%.

tasks.py: stdlib-only cross-platform task runner. Targets:
  lint / format / format-fix / typecheck / test / run /
  docker-build / docker-run / ci (= lint + format-check +
  typecheck + test). Forwards extra args after `--` to run /
  docker-run.

Stage 6 exit criterion verified: `python tasks.py ci` runs every
gate and reports "All gates passed." with 127 tests green and
96.60% coverage.

+++ src/pick_at_random/infrastructure/ntp_time_source.py
@@ -28,8 +28,7 @@ class NtpFetchError(RuntimeError):
 class _NtpClient(Protocol):
     def request(
         self, host: str, version: int = ..., port: str = ..., timeout: float = ...
-    ) -> ntplib.NTPStats:
-        ...
+    ) -> ntplib.NTPStats: ...
 
 
 class NtpTimeSource:
@@ -48,9 +47,7 @@ class NtpTimeSource:
         if version not in (3, 4):
             raise ValueError(f"version must be 3 or 4, got {version}.")
         if timeout_seconds <= 0:
-            raise ValueError(
-                f"timeout_seconds must be positive, got {timeout_seconds}."
-            )
+            raise ValueError(f"timeout_seconds must be positive, got {timeout_seconds}.")
         self._server = server
         self._version = version
         self._timeout = timeout_seconds
@@ -62,21 +59,14 @@ class NtpTimeSource:
                 self._server, version=self._version, timeout=self._timeout
             )
         except (ntplib.NTPException, OSError) as exc:
-            raise NtpFetchError(
-                f"NTP query failed for {self._server!r}: {exc}"
-            ) from exc
+            raise NtpFetchError(f"NTP query failed for {self._server!r}: {exc}") from exc
 
         tx_time = float(response.tx_time)
         if tx_time <= 0:
-            raise NtpFetchError(
-                f"NTP server {self._server!r} returned non-positive timestamp."
-            )
+            raise NtpFetchError(f"NTP server {self._server!r} returned non-positive timestamp.")
 
-        unix_nanoseconds = int(round(tx_time * 1_000_000_000))
-        iso_timestamp = (
-            datetime.fromtimestamp(tx_time, tz=UTC)
-            .isoformat(timespec="microseconds")
-        )
+        unix_nanoseconds = round(tx_time * 1_000_000_000)
+        iso_timestamp = datetime.fromtimestamp(tx_time, tz=UTC).isoformat(timespec="microseconds")
         return NtpDraw(
             server=self._server,
             unix_nanoseconds=unix_nanoseconds,
